Memperbaiki tampilan tombol kemablai dan selnajutnya di form lihat, update dan tambah data siswa

+++ src/Pages/Admin/lihat-data/Biodata.jsx
@@ -61,12 +61,23 @@ const Biodata = () => {
   }, [id]);
 
   const backButton = () => {
-    const token = localStorage.getItem("token"); // Simpan token sebelum clear
-    localStorage.clear(); // Hapus semua data
-    if (token) {
-      localStorage.setItem("token", token); // Simpan kembali token
-    }
-    navigate("/admin/dashboard");
+    Swal.fire({
+      icon: 'warning',
+      title: 'Peringatan!',
+      text: 'Anda akan dialihkan ke halaman data siswa. Apakah Anda yakin ingin melanjutkan?',
+      showCancelButton: true,
+      confirmButtonText: 'Lanjutkan',
+      cancelButtonText: 'Batal',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        const token = localStorage.getItem("token"); // Simpan token sebelum clear
+        localStorage.clear(); // Hapus semua data
+        if (token) {
+          localStorage.setItem("token", token); // Simpan kembali token
+        }
+        navigate(`/admin/datasiswa`);
+      }
+    });
   };
 
   const nextButton = () => {
@@ -163,7 +174,7 @@ const Biodata = () => {
   if (error) return <p>{error}</p>;
 
   return (
-    <div className="bg-gray-100 w-screen px-10 pb-6 h-screen overflow-y-auto text-xl">
+    <div className="bg-gray-100 w-screen px-10 pb-6 h-screen overflow-y-auto text-xl justify-end">
       {/* Profil dan Input Halaman */}
       <div className="my-10 w-full flex flex-col gap-6">
         <Profil />
@@ -267,7 +278,7 @@ const Biodata = () => {
         </div>
       </div>
       {/* Tombol Next & Back */}
-      <div className="grid grid-cols-2 space-x-4">
+      <div className="flex justify-end space-x-4">
         <Nextbefore next={nextButton} back={backButton} />
       </div>
     </div>
